+++ src/demo/FakeAgrDataService.js
@@ -1,32 +1,5 @@
-export class FakeDataService{
+export class FakeAgrDataService{
 
-  // Our variant endpoint should take a range on a chromosome
-  // and return all the variants that fall within that range
-  GetFakeVariants(chromosome, start, end){
-    let fakeVariants = [
-      {
-        "position": 48515449,
-        "ref": "G",
-        "mutant":"A"
-      },
-      {
-        "position": 48515458,
-        "ref": "T",
-        "mutant":"C"
-      },
-      {
-        "position": 48515461,
-        "ref": "A",
-        "mutant":"G"
-      },
-      {
-        "position": 48515500,
-        "ref": "C",
-        "mutant":"A"
-      }
-    ];
-    return fakeVariants;
-  }
 
   GetFakeFishGeneDataMyl7(){
 // http://localhost:8080/apollo/track/remotefish2/All%20Genes/8:40454020..40474021.json?includeGenotypes=false&ignoreCache=true
@@ -69,31 +42,5 @@ export class FakeDataService{
     // return [{"symbol":{"meta":{"number":[1],"description":["The human readable name of the allele"],"id":["symbol"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"tm6709\""]},"alternative_alleles":{"meta":{"description":"VCF ALT field, list of alternate non-reference alleles called on at least one of the samples"},"values":"A"},"geneLevelConsequence":{"meta":{"number":[-1],"description":["VEP consequence of the variant"],"id":["geneLevelConsequence"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":[["\"splice_acceptor_variant","splice_donor_variant","stop_lost","inframe_deletion","intron_variant\""]]},"hgvs_nomenclature":{"meta":{"number":[1],"description":["the HGVS name of the allele"],"id":["hgvs_nomenclature"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"NC_003283.11:g.7355_7776del\""]},"impact":{"meta":{"number":[1],"description":["Variant impact scale"],"id":["impact"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"HIGH\""]},"description":"deletion ACCAGTAGTGATCGAGCAATTGCGTTCTTCTCTGAAGCTTCTATGATTCATTGAATAAAATATATTTCTCAAAACGTACTTGCTTATCGACAACAACCAACCAACGTCCACCTTGAACGTTGTTGACGTCCTCCCACATTGGCTTGATTCCTTCCTTGAACAAGTAATAATCGGATCCCCAGTTCAATCCTCCGGCAGACTGAATGTGATTGTACAGCGACCAGAAGTCCTCGACAGTGTCGAAAAGTGAAACCATCTGGAAAAAATCGATAAAAGACGTATTTAAAAATCTTCTACCTTCAGACAATCCTCCCATTCCTTGTTACGGTCAGCTTTCAAGTACCAGAGAGCCCAGCGATTCTGGAGGGGGTGTCTGGTGAGAAGCTCTGGAGGAACTGAAGCATCGGACGCATTCACATCGCC > A","type":"deletion","symbol_text":{"meta":{"number":[1],"description":["Another human readable representation of the allele"],"id":["symbol_text"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"tm6709\""]},"fmax":7776,"reference_allele":"ACCAGTAGTGATCGAGCAATTGCGTTCTTCTCTGAAGCTTCTATGATTCATTGAATAAAATATATTTCTCAAAACGTACTTGCTTATCGACAACAACCAACCAACGTCCACCTTGAACGTTGTTGACGTCCTCCCACATTGGCTTGATTCCTTCCTTGAACAAGTAATAATCGGATCCCCAGTTCAATCCTCCGGCAGACTGAATGTGATTGTACAGCGACCAGAAGTCCTCGACAGTGTCGAAAAGTGAAACCATCTGGAAAAAATCGATAAAAGACGTATTTAAAAATCTTCTACCTTCAGACAATCCTCCCATTCCTTGTTACGGTCAGCTTTCAAGTACCAGAGAGCCCAGCGATTCTGGAGGGGGTGTCTGGTGAGAAGCTCTGGAGGAACTGAAGCATCGGACGCATTCACATCGCC","name":"NC_003283.11:g.7355_7776del","allele_of_genes":{"meta":{"number":[1],"description":["The genes that the Allele is located on"],"id":["allele_of_genes"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"WB:WBGene00002061\""]},"fmin":7353,"seqId":"V","uniqueID":"NC_003283.11:g.7355_7776del","alleles":{"meta":{"number":[-1],"description":["The alleles of the variant"],"id":["alleles"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"WB:WBVar02125540\""]}},{"symbol":{"meta":{"number":[1],"description":["The human readable name of the allele"],"id":["symbol"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"tm774\""]},"alternative_alleles":{"meta":{"description":"VCF ALT field, list of alternate non-reference alleles called on at least one of the samples"},"values":"C"},"geneLevelConsequence":{"meta":{"number":[-1],"description":["VEP consequence of the variant"],"id":["geneLevelConsequence"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"intron_variant\""]},"hgvs_nomenclature":{"meta":{"number":[1],"description":["the HGVS name of the allele"],"id":["hgvs_nomenclature"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"NC_003283.11:g.23999_24591del\""]},"impact":{"meta":{"number":[1],"description":["Variant impact scale"],"id":["impact"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"MODIFIER\""]},"description":"deletion CTAACAGAAAATGTACTATTACACCGATTATTAGTACTTATAATTTTTTTTTTGAAAGTTGGAAAATATTTTAAAGTTTGTAAAAATTCTATAAATAATATTGAATATTATAACGGTAAAATATAATATTTATAGAATTTTTTCAATTTTCAAAAAAGTACTGTTTGTCATTTGTTTTTGTTTGTTAGAATACCTTTTAATAGCGGTACCGAAATCTGGGAAATATTTTCAAATGACTCCAAATTTTCGCCTGATTCCGAATATCTATGTGAAAAAAAGTGAAAAAAAACATCCCTTATTTTATATTTCATTTCCAATTCCCAATTTCATTTGTGCCCGCATTACTTTTTTCAAACGCGCGCCCAAATAAATTTTCATTAGAGCGCGTTTGCTTTGTGTCGATTTACGGGAGCTCTCCATTTACAAAAAAAATTAACGCTTTTTTTAAACAGTTTTTAAGCGAGTTTCCTACTTTTTTAATCGATTTCAGCGGAACGGTTTTTCTTGAAATATCGTTTTTAAGGTCAATTGATATGATAATAAAATATATTGTTGACCTTAAAACACATTAAAAAAACTTCGACAAGAACCGAC > C","type":"deletion","symbol_text":{"meta":{"number":[1],"description":["Another human readable representation of the allele"],"id":["symbol_text"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"tm774\""]},"fmax":24591,"reference_allele":"CTAACAGAAAATGTACTATTACACCGATTATTAGTACTTATAATTTTTTTTTTGAAAGTTGGAAAATATTTTAAAGTTTGTAAAAATTCTATAAATAATATTGAATATTATAACGGTAAAATATAATATTTATAGAATTTTTTCAATTTTCAAAAAAGTACTGTTTGTCATTTGTTTTTGTTTGTTAGAATACCTTTTAATAGCGGTACCGAAATCTGGGAAATATTTTCAAATGACTCCAAATTTTCGCCTGATTCCGAATATCTATGTGAAAAAAAGTGAAAAAAAACATCCCTTATTTTATATTTCATTTCCAATTCCCAATTTCATTTGTGCCCGCATTACTTTTTTCAAACGCGCGCCCAAATAAATTTTCATTAGAGCGCGTTTGCTTTGTGTCGATTTACGGGAGCTCTCCATTTACAAAAAAAATTAACGCTTTTTTTAAACAGTTTTTAAGCGAGTTTCCTACTTTTTTAATCGATTTCAGCGGAACGGTTTTTCTTGAAATATCGTTTTTAAGGTCAATTGATATGATAATAAAATATATTGTTGACCTTAAAACACATTAAAAAAACTTCGACAAGAACCGAC","name":"NC_003283.11:g.23999_24591del","allele_of_genes":{"meta":{"number":[1],"description":["The genes that the Allele is located on"],"id":["allele_of_genes"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"WB:WBGene00001177\""]},"fmin":23997,"seqId":"V","uniqueID":"NC_003283.11:g.23999_24591del","alleles":{"meta":{"number":[-1],"description":["The alleles of the variant"],"id":["alleles"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"WB:WBVar00249802\""]}},{"symbol":{"meta":{"number":[1],"description":["The human readable name of the allele"],"id":["symbol"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"nj77\""]},"alternative_alleles":{"meta":{"description":"VCF ALT field, list of alternate non-reference alleles called on at least one of the samples"},"values":"A"},"geneLevelConsequence":{"meta":{"number":[-1],"description":["VEP consequence of the variant"],"id":["geneLevelConsequence"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"splice_donor_variant\""]},"hgvs_nomenclature":{"meta":{"number":[1],"description":["the HGVS name of the allele"],"id":["hgvs_nomenclature"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"NC_003283.11:g.26381G>A\""]},"impact":{"meta":{"number":[1],"description":["Variant impact scale"],"id":["impact"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"HIGH\""]},"description":"SNV G > A","type":"SNV","symbol_text":{"meta":{"number":[1],"description":["Another human readable representation of the allele"],"id":["symbol_text"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"nj77\""]},"fmax":26381,"reference_allele":"G","name":"NC_003283.11:g.26381G>A","allele_of_genes":{"meta":{"number":[1],"description":["The genes that the Allele is located on"],"id":["allele_of_genes"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"WB:WBGene00001177\""]},"fmin":26380,"seqId":"V","uniqueID":"NC_003283.11:g.26381G>A","alleles":{"meta":{"number":[-1],"description":["The alleles of the variant"],"id":["alleles"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"WB:WBVar00604105\""]}},{"symbol":{"meta":{"number":[1],"description":["The human readable name of the allele"],"id":["symbol"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"n488\""]},"alternative_alleles":{"meta":{"description":"VCF ALT field, list of alternate non-reference alleles called on at least one of the samples"},"values":"T"},"geneLevelConsequence":{"meta":{"number":[-1],"description":["VEP consequence of the variant"],"id":["geneLevelConsequence"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":[["\"splice_acceptor_variant","splice_donor_variant","coding_sequence_variant","intron_variant\""]]},"hgvs_nomenclature":{"meta":{"number":[1],"description":["the HGVS name of the allele"],"id":["hgvs_nomenclature"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"NC_003283.11:g.28912_30730del\""]},"impact":{"meta":{"number":[1],"description":["Variant impact scale"],"id":["impact"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"HIGH\""]},"description":"deletion TAACTTAATAAGTTATGAAAGCAAGAATAAATAGAATGACAGGTTGCTTGAAACGATGGAATGCGATATTTTCTCTTATAGGCTTCTTCTTTTTAGGATTTTTTATAAAAAATGAGGTGTCGGCTGCTACCTGTGAGGTTTTGGCTGCCACCTATAAAGTGTCGGCTGCTACCCATGAGGTGTCCGCTGCTACCCATGAGGTGCCGGCTGCATAGTCTCACAGGTTTAGTGGTTTTCTTGCACGAAATATACTATTTTTGAAAGAAACAGTAAACCCATTCGTGCCTTCATCGTTTCTCAAACTGTCATTTTATGGGTCTTGCCACGAAACTTCTCCTCTCTATCACTCTACGTATCTTTATCTGTCTATAATATTTGGCCGCCCTAGTCTCCCCCTTCCGTCACTGTCTCTCTATGTATTTCAGCAGGCGATCAGCGTGCTCATCCGGAAGTCGAGCAACCCGTGAGCTCATCCTCTCCAGCCACACCGTCCATCAGCGGACCCCCGCCGTGTGCCACGTCATCAGGTTCAACGTCATCAATCACTATTACCACAACGGGATGCTCCACGTCATCGTCGGGACCCTCAAAACATATTCTAGGGGGTAAGAATTAGTTGTTTGCCCTGCGATCCTGAATCAAAATAATTGATATTTCTCTACCCCCGGTAGCTTCTTTTTTTTCGTACTTCCCTCTTCCCCTAACCAACAGAAACGTGCGCTTTTTTCTCCTTTCCTTCTATTTTTCTTTCCATTTCTTTTCTCTTTGGCTTGCCGCGGGCTGCAGCTTTTTTTGAACCACGATTTTTGAATTGATTTATTTTTAATTTTTTTTGGTTTGCAAATTTGGCTAGAAAATATGAAAATTGGAAAATAGTTTTAGAGGAATTTTGTTCAAAAATGTAGGTTAAGATACATTGATACTAAATCCAGGGAAGATATCTAGCGAACGAATCTCATTAGGTGTCGGCTGCTACATGTCGGCTGCACCCCAAAATGTGACGGTTATATACTGTCGGCTGCAACCCATAAGGTGTCGGCTGCAACCCACAAGGTGTCGGCTGCAACCTACAAGGTGTCGGCTGCAACCCACGAATTTTCGGCAGCAACCCACGGGGTGTCGGCTGCCACCCACAAGTTGTCGGCTGTAACCCACAAGGTGTCAGCTGCAACCTACGAAGTGTCGGCTGCTACTTGTCGGTTGCAACCCTGCCCTCACAAAAGCAGCCGGCGTCCAGCGGTTCCCTTATGATTTCTTCAAAAAAAACCGTAAAACTTTCAATTTTTTAGGTGAAATGCCCGCCAAGGAAAACGACGAAGCACATCCGGAGCTAAAGCAGAATTTCATTGCTAAAAATTTGAAAGGATTCGGGTTTTCGAAAAAACAAGTGAGTTTTTTTGAAATTTTCATTTTTTTTCTTCTTTTTCAGATTTCATTATTTCAGGAAACAAACAAAATTTCAAATTTTAACATAGTTTTTAAGGGTTTCCTCTTCCTCCTTATTCTTATTCTTCTCTCTCAAATCCCTTAACTTTTTTATGCATTTGCGCTCTACTGATAATCGAAATTCTCATAAGAGCGTATTTGCACTAAACTTGCCTGCGTTTGAAGTGTTCCTCTTCCTAAGCTGTCTCATAATTTCTGTTCTCTGTTGATTATCATTTTCTTGTTTTTCTGTTTCGTTTCTCTGATTTTCTCGTAGAAAACTGTTGTTTTTTTTCTGATTTCCATGTACTCTTACAACGTACTACTAACTCTCACTTGTTGGGTAGTAGTGCAGATTAAACAAAAAATAAAATCTGAAAGTTTGTAATTTCAGA > T","type":"deletion","symbol_text":{"meta":{"number":[1],"description":["Another human readable representation of the allele"],"id":["symbol_text"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"n488\""]},"fmax":30730,"reference_allele":"TAACTTAATAAGTTATGAAAGCAAGAATAAATAGAATGACAGGTTGCTTGAAACGATGGAATGCGATATTTTCTCTTATAGGCTTCTTCTTTTTAGGATTTTTTATAAAAAATGAGGTGTCGGCTGCTACCTGTGAGGTTTTGGCTGCCACCTATAAAGTGTCGGCTGCTACCCATGAGGTGTCCGCTGCTACCCATGAGGTGCCGGCTGCATAGTCTCACAGGTTTAGTGGTTTTCTTGCACGAAATATACTATTTTTGAAAGAAACAGTAAACCCATTCGTGCCTTCATCGTTTCTCAAACTGTCATTTTATGGGTCTTGCCACGAAACTTCTCCTCTCTATCACTCTACGTATCTTTATCTGTCTATAATATTTGGCCGCCCTAGTCTCCCCCTTCCGTCACTGTCTCTCTATGTATTTCAGCAGGCGATCAGCGTGCTCATCCGGAAGTCGAGCAACCCGTGAGCTCATCCTCTCCAGCCACACCGTCCATCAGCGGACCCCCGCCGTGTGCCACGTCATCAGGTTCAACGTCATCAATCACTATTACCACAACGGGATGCTCCACGTCATCGTCGGGACCCTCAAAACATATTCTAGGGGGTAAGAATTAGTTGTTTGCCCTGCGATCCTGAATCAAAATAATTGATATTTCTCTACCCCCGGTAGCTTCTTTTTTTTCGTACTTCCCTCTTCCCCTAACCAACAGAAACGTGCGCTTTTTTCTCCTTTCCTTCTATTTTTCTTTCCATTTCTTTTCTCTTTGGCTTGCCGCGGGCTGCAGCTTTTTTTGAACCACGATTTTTGAATTGATTTATTTTTAATTTTTTTTGGTTTGCAAATTTGGCTAGAAAATATGAAAATTGGAAAATAGTTTTAGAGGAATTTTGTTCAAAAATGTAGGTTAAGATACATTGATACTAAATCCAGGGAAGATATCTAGCGAACGAATCTCATTAGGTGTCGGCTGCTACATGTCGGCTGCACCCCAAAATGTGACGGTTATATACTGTCGGCTGCAACCCATAAGGTGTCGGCTGCAACCCACAAGGTGTCGGCTGCAACCTACAAGGTGTCGGCTGCAACCCACGAATTTTCGGCAGCAACCCACGGGGTGTCGGCTGCCACCCACAAGTTGTCGGCTGTAACCCACAAGGTGTCAGCTGCAACCTACGAAGTGTCGGCTGCTACTTGTCGGTTGCAACCCTGCCCTCACAAAAGCAGCCGGCGTCCAGCGGTTCCCTTATGATTTCTTCAAAAAAAACCGTAAAACTTTCAATTTTTTAGGTGAAATGCCCGCCAAGGAAAACGACGAAGCACATCCGGAGCTAAAGCAGAATTTCATTGCTAAAAATTTGAAAGGATTCGGGTTTTCGAAAAAACAAGTGAGTTTTTTTGAAATTTTCATTTTTTTTCTTCTTTTTCAGATTTCATTATTTCAGGAAACAAACAAAATTTCAAATTTTAACATAGTTTTTAAGGGTTTCCTCTTCCTCCTTATTCTTATTCTTCTCTCTCAAATCCCTTAACTTTTTTATGCATTTGCGCTCTACTGATAATCGAAATTCTCATAAGAGCGTATTTGCACTAAACTTGCCTGCGTTTGAAGTGTTCCTCTTCCTAAGCTGTCTCATAATTTCTGTTCTCTGTTGATTATCATTTTCTTGTTTTTCTGTTTCGTTTCTCTGATTTTCTCGTAGAAAACTGTTGTTTTTTTTCTGATTTCCATGTACTCTTACAACGTACTACTAACTCTCACTTGTTGGGTAGTAGTGCAGATTAAACAAAAAATAAAATCTGAAAGTTTGTAATTTCAGA","name":"NC_003283.11:g.28912_30730del","allele_of_genes":{"meta":{"number":[1],"description":["The genes that the Allele is located on"],"id":["allele_of_genes"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"INTEGER"}]},"values":["\"WB:WBGene00001177\""]},"fmin":28910,"seqId":"V","uniqueID":"NC_003283.11:g.28912_30730del","alleles":{"meta":{"number":[-1],"description":["The alleles of the variant"],"id":["alleles"],"type":[{"enumType":"htsjdk.variant.vcf.VCFHeaderLineCount","name":"UNBOUNDED"}]},"values":["\"WB:WBVar00089535\""]}}];
   }
 
-  // Fake globll variant track
-  GetFakeGlobalVariants(chromosome, start, end){
-    let fakeVariants = [
-      {
-        "position": 75575916,
-        "ref": "G",
-        "mutant":"A"
-      },
-      {
-        "position": 75575056,
-        "ref": "T",
-        "mutant":"C"
-      },
-      {
-        "position": 75645056,
-        "ref": "A",
-        "mutant":"G"
-      },
-      {
-        "position": 75655056,
-        "ref": "C",
-        "mutant":"A"
-      }
-    ]
-    return fakeVariants;
-  }
 
 }
