+++ src/tracks/VariantTrackGlobal.js
@@ -72,7 +72,6 @@ export default class VariantTrack {
     /* Method to get reference label */
     async getTrackData()
     {
-        let apolloService = new ApolloService()
-        this.variants =  await apolloService.GetFakeGlobalVariants();
+        this.variants =  await new ApolloService().GetFakeGlobalVariants();
     }
 }
